Fixed auto actions and added more readability to action creation
Fixed problems with auto actions relating to drive path not running with the correct timeframe, relating to drive path not locking its own thread lock, and relating to parallel action and series action potentially trying to run actions that have the same lock as them, which could potentially stop the entire parallel action and/or series action. Added readability relating to the will cancel parameter in the action constructor via creating two constants in the Constants file under Action Constants that gives names to the two different possibilities for this parameter, allowing people reading the code to read WILL_CANCEL and/or WILL_NOT_CANCEL instead of true and/or false.